requirements.txt
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==3.0.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
comm==0.2.2
debugpy==1.8.13
decorator==5.2.1
dnspython==2.7.0
email-validator==2.2.0
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.11
fastapi-cli==0.0.7
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib-metadata==8.6.1
ipykernel==6.29.5
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.6
jupyter-client==8.6.3
jupyter-core==5.7.2
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mysql-connector-python==9.2.0
nest-asyncio==1.6.0
orjson==3.10.16
packaging==24.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.7
prompt-toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-extra-types==2.10.3
pydantic-settings==2.9.1
pygments==2.19.1
pyjwt==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pyyaml==6.0.2
pyzmq==26.3.0
rich==13.9.4
rich-toolkit==0.13.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.40
sqlmodel==0.0.24
stack-data==0.6.3
starlette==0.46.1
tornado==6.4.2
traitlets==5.14.3
typer==0.15.2
typing-extensions==4.12.2
typing-inspection==0.4.0
ujson==5.10.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0.1
zipp==3.21.0
